move OptimizedNautobotObjectType to nautobot core
