+++ nautobot/circuits/graphql/types.py
@@ -2,9 +2,9 @@ from graphene import Field
 
 from nautobot.circuits.filters import CircuitTerminationFilterSet
 from nautobot.circuits.models import CircuitTermination
+from nautobot.core.graphql.types import OptimizedNautobotObjectType
 from nautobot.core.graphql.utils import construct_resolver
 from nautobot.dcim.graphql.mixins import CableTerminationMixin, PathEndpointMixin
-from nautobot.extras.graphql.types import OptimizedNautobotObjectType
 
 
 class CircuitTerminationType(OptimizedNautobotObjectType, CableTerminationMixin, PathEndpointMixin):
