+++ nautobot/extras/graphql/types.py
@@ -1,29 +1,6 @@
-import graphene
-from graphene_django import DjangoObjectType
-import graphene_django_optimizer as gql_optimizer
-
+from nautobot.core.graphql.types import OptimizedNautobotObjectType
 from nautobot.extras.models import Status, Tag, DynamicGroup
 from nautobot.extras.filters import StatusFilterSet, TagFilterSet, DynamicGroupFilterSet
-
-
-class NautobotObjectType(DjangoObjectType):
-    url = graphene.String()
-
-    def resolve_url(self, args):
-        return self.get_absolute_url()
-
-    class Meta:
-        abstract = True
-
-
-class OptimizedNautobotObjectType(gql_optimizer.OptimizedDjangoObjectType):
-    url = graphene.String()
-
-    def resolve_url(self, args):
-        return self.get_absolute_url()
-
-    class Meta:
-        abstract = True
 
 
 class TagType(OptimizedNautobotObjectType):
